+++ ui/src/index.js
@@ -37,6 +37,8 @@ import QasSingleView from './components/single-view/QasSingleView'
 import QasSortable from './components/sortable/QasSortable'
 import QasTableGenerator from './components/table-generator/QasTableGenerator'
 import QasTabsGenerator from './components/tabs-generator/QasTabsGenerator'
+import QasTextTruncate from './components/text-truncate/QasTextTruncate'
+import QasTip from './components/tip/QasTip'
 import QasTransfer from './components/transfer/QasTransfer'
 import QasUploader from './components/uploader/QasUploader'
 
@@ -100,6 +102,8 @@ export {
   QasSortable,
   QasTableGenerator,
   QasTabsGenerator,
+  QasTextTruncate,
+  QasTip,
   QasTransfer,
   QasUploader,
 
@@ -149,6 +153,8 @@ export default {
   QasSortable,
   QasTableGenerator,
   QasTabsGenerator,
+  QasTextTruncate,
+  QasTip,
   QasTransfer,
   QasUploader,
 
@@ -197,6 +203,8 @@ export default {
     Vue.component('QasSortable', QasSortable)
     Vue.component('QasTableGenerator', QasTableGenerator)
     Vue.component('QasTabsGenerator', QasTabsGenerator)
+    Vue.component('QasTextTruncate', QasTextTruncate)
+    Vue.component('QasTip', QasTip)
     Vue.component('QasTransfer', QasTransfer)
     Vue.component('QasUploader', QasUploader)